data_loader/requirements.data_loader.txt
geopandas
psycopg2-binary
sqlalchemy
requests
beautifulsoup4
python-dotenv==1.0.1
geoalchemy2==0.17.1
Jinja2==3.1.5
PyYAML